Extract formatting logic from TextUi to a Formatter class
